Only check at sample points in at most one shadow map
This was intended as an optimisation, but it ended up making things slower.
The approach was inspired by how OSG culls subgraphs with polytopes, keeping a mask so it doesn't recheck once it's determined that a parent node was entirely on one side of a plane.

As the texture accesses seem to be by far the most expensive part of shadow coverage calculation, I'd hoped that avoiding some accesses would help, but apparently not - the extra maths in a tight loop ends up costing more than it saves on average.
There might be a bit more performance to eek out of this if I used the GLSL extension that makes 64-bit integer types available, but as it actively hurt performance rather than just not helping as much as I'd hoped, I don't think it would actually be enough to even break even.

Thinking about it, the reason this doesn't help is that even though it skips some texture accesses, it doesn't do anything to decrease divergence.