Add options to improve cache hit rate
--qs sorts queries in one axis, and --qb bins queries in another axis.
As they operate in different axes, they can be used in conjunction, and for large datasets with many queries, this gives the best throughput.

With the simple implementations here, these optimisations don't necessarily pay for themselves - they take time to apply, and don't necessarily save a larger amount of time.
An application that has an alternative way to get a similar effect, e.g. splitting the simulation world into cells, and processing all entities within a cell before moving onto the next cell, may not have to pay the same costs to get the same benefit.